+++ src/lib/litegraph/src/canvas/ToInputFromIoNodeLink.ts
@@ -1,5 +1,4 @@
 import type { LGraphNode } from '@/lib/litegraph/src/LGraphNode'
-import { LiteGraph } from '@/lib/litegraph/src/litegraph'
 import type { LLink } from '@/lib/litegraph/src/LLink'
 import type { Reroute } from '@/lib/litegraph/src/Reroute'
 import type { CustomEventTarget } from '@/lib/litegraph/src/infrastructure/CustomEventTarget'
@@ -26,7 +25,7 @@ export class ToInputFromIoNodeLink implements RenderLink {
   fromDirection: LinkDirection = LinkDirection.RIGHT
   readonly existingLink?: LLink
   disconnectOnDrop: boolean
-  readonly disconnectOrigin: Point | undefined
+  readonly disconnectOrigin?: Point
   readonly isIoNodeLink = true
 
   constructor(
@@ -169,23 +168,5 @@ export class ToInputFromIoNodeLink implements RenderLink {
       })
 
     return true
-  }
-  drawConnectionCircle(ctx: CanvasRenderingContext2D, to: Readonly<Point>) {
-    if (!this.disconnectOnDrop || !this.disconnectOrigin) return
-
-    const [originX, originY] = this.disconnectOrigin
-    const radius = 35
-    const distSquared = (originX - to[0]) ** 2 + (originY - to[1]) ** 2
-
-    ctx.save()
-    ctx.strokeStyle = LiteGraph.WIDGET_OUTLINE_COLOR
-    ctx.lineWidth = 2
-    ctx.beginPath()
-    ctx.moveTo(originX + radius, originY)
-    ctx.arc(originX, originY, radius, 0, Math.PI * 2)
-    ctx.stroke()
-    ctx.restore()
-
-    this.disconnectOnDrop = distSquared < radius ** 2
   }
 }
